버그 수정
플레이어
 - 힐 이동 제어 길이 수정(애니메이션 길이만큼)
 - 피격 후 무적 시간 증가
 - 피격 사운드 추가
 - 패링 후 피격 수정
 - 피격 후 패링 수정
 - 공중 힐 게이지 차감 수정

무신
 - 각 공격 콜라이더 크기 수정

길달
 - 체력바 구현
 - 연출 중 카메라 흔들림 추가

시나리오
 - 선택씬 카메라 흔들림 복구
 - 추락씬 텍스트 위치 수정

시스템
 - 첫 플레이 이후 메인화면 isFirst 수정
 - 시연용 무적 기능 추가(F1 버튼)

+++ Assets/Scripts/Boss/DokkaebiWave.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class DokkaebiWave : MonoBehaviour
 {
     [Header("Projectile Setting")]
-    public int fireTime = 2;
+    public float fireTime = 2;
     [SerializeField] private int baseDamage = 1;
 
     private Animator anim;
